gpu: nvgpu: Don't cast runlist size to u64
We multiply integer entry size and number of runlist entries and
store the result in u64. The result is used as size of memory, so
it should be size_t instead.

Reviewed-on: http://git-master/r/1249957
GVS: Gerrit_Virtual_Submit

+++ drivers/gpu/nvgpu/gk20a/fifo_gk20a.c
@@ -590,7 +590,7 @@ static int init_runlist(struct gk20a *g, struct fifo_gk20a *f)
 	struct device *d = dev_from_gk20a(g);
 	s32 runlist_id = -1;
 	u32 i;
-	u64 runlist_size;
+	size_t runlist_size;
 
 	gk20a_dbg_fn("");
 
@@ -619,7 +619,7 @@ static int init_runlist(struct gk20a *g, struct fifo_gk20a *f)
 			goto clean_up_runlist;
 
 		runlist_size  = f->runlist_entry_size * f->num_runlist_entries;
-		gk20a_dbg_info("runlist_entries %d runlist size %llu\n",
+		gk20a_dbg_info("runlist_entries %d runlist size %zu\n",
 					f->num_runlist_entries, runlist_size);
 
 		for (i = 0; i < MAX_RUNLIST_BUFFERS; i++) {
